+++ apps/block_scout_web/lib/block_scout_web/views/tokens_view.ex
@@ -1,0 +1,8 @@
+defmodule BlockScoutWeb.TokensView do
+  use BlockScoutWeb, :view
+
+  alias Explorer.Chain.Token
+
+  def decimals?(%Token{decimals: nil}), do: false
+  def decimals?(%Token{decimals: _}), do: true
+end