+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Current
 
 ### Features
+- [#3227](https://github.com/poanetwork/blockscout/pull/3227) - Distinguishing of bridged tokens
+- [#3224](https://github.com/poanetwork/blockscout/pull/3224) - Top tokens page
+
+### Fixes
+- [#3226](https://github.com/poanetwork/blockscout/pull/3226) - Fix notifier query for live update of token transfers
+- [#3220](https://github.com/poanetwork/blockscout/pull/3220) - Allow interaction with navbar menu at block-not-found page
+
+### Chore
+
+
+## 3.3.1-beta
+
+### Features
 - [#3216](https://github.com/poanetwork/blockscout/pull/3216) - Display new token transfers at token page and address page without refreshing the page
 - [#3199](https://github.com/poanetwork/blockscout/pull/3199) - Show compilation error at contract verification
 - [#3193](https://github.com/poanetwork/blockscout/pull/3193) - Raw trace copy button